Refactoring improvements
