+++ src/PSRule.Rules.Azure/Data/Template/BaseTemplateContext.cs
@@ -1,15 +1,14 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace PSRule.Rules.Azure.Data.Template
-{
-    /// <summary>
-    /// A base implementation for template contexts.
-    /// </summary>
-    internal abstract class BaseTemplateContext
-    {
-        public virtual bool ShouldThrowMissingProperty => true;
+namespace PSRule.Rules.Azure.Data.Template;
 
-        public DebugSymbol DebugSymbol { get; set; }
-    }
+/// <summary>
+/// A base implementation for template contexts.
+/// </summary>
+internal abstract class ResourceManagerVisitorContext
+{
+    public virtual bool ShouldThrowMissingProperty => true;
+
+    public DebugSymbol DebugSymbol { get; set; }
 }
