+++ src/PSRule.Rules.Azure/Data/Template/DeploymentValue.cs
@@ -7,132 +7,131 @@ using System.Diagnostics;
 using Newtonsoft.Json.Linq;
 using static PSRule.Rules.Azure.Data.Template.TemplateVisitor;
 
-namespace PSRule.Rules.Azure.Data.Template
+namespace PSRule.Rules.Azure.Data.Template;
+
+[DebuggerDisplay("{Id}")]
+internal sealed class DeploymentValue : BaseResourceValue, IResourceValue, ILazyObject
 {
-    [DebuggerDisplay("{Id}")]
-    internal sealed class DeploymentValue : BaseResourceValue, IResourceValue, ILazyObject
+    private const string RESOURCE_TYPE_DEPLOYMENT = "Microsoft.Resources/deployments";
+
+    private const string PROPERTY_PROPERTIES = "properties";
+    private const string PROPERTY_OUTPUTS = "outputs";
+
+    private readonly Lazy<JObject> _Value;
+    private readonly Dictionary<string, ILazyValue> _Outputs;
+
+    internal DeploymentValue(string id, string name, string symbolicName, string scope, DeploymentScope deploymentScope, JObject value, TemplateContext.CopyIndexState copy)
+        : this(id, name, symbolicName, scope, deploymentScope, () => value, copy) { }
+
+    internal DeploymentValue(string id, string name, string symbolicName, string scope, DeploymentScope deploymentScope, Func<JObject> value, TemplateContext.CopyIndexState copy)
+        : base(id, name, symbolicName)
     {
-        private const string RESOURCETYPE_DEPLOYMENT = "Microsoft.Resources/deployments";
+        _Value = new Lazy<JObject>(value);
+        Copy = copy;
+        Properties = new DeploymentProperties(this);
+        _Outputs = new Dictionary<string, ILazyValue>(StringComparer.OrdinalIgnoreCase);
+        Scope = scope;
+        DeploymentScope = deploymentScope;
+    }
 
-        private const string PROPERTY_PROPERTIES = "properties";
-        private const string PROPERTY_OUTPUTS = "outputs";
+    private sealed class DeploymentProperties : ILazyObject
+    {
+        private readonly DeploymentValue _Deployment;
+        private readonly JObject _Properties;
+        private readonly DeploymentOutputs _Outputs;
 
-        private readonly Lazy<JObject> _Value;
-        private readonly Dictionary<string, ILazyValue> _Outputs;
-
-        internal DeploymentValue(string id, string name, string symbolicName, string scope, DeploymentScope deploymentScope, JObject value, TemplateContext.CopyIndexState copy)
-            : this(id, name, symbolicName, scope, deploymentScope, () => value, copy) { }
-
-        internal DeploymentValue(string id, string name, string symbolicName, string scope, DeploymentScope deploymentScope, Func<JObject> value, TemplateContext.CopyIndexState copy)
-            : base(id, name, symbolicName)
+        public DeploymentProperties(DeploymentValue deployment)
         {
-            _Value = new Lazy<JObject>(value);
-            Copy = copy;
-            Properties = new DeploymentProperties(this);
-            _Outputs = new Dictionary<string, ILazyValue>(StringComparer.OrdinalIgnoreCase);
-            Scope = scope;
-            DeploymentScope = deploymentScope;
+            _Deployment = deployment;
+            _Deployment.Value.TryGetProperty(PROPERTY_PROPERTIES, out _Properties);
+            _Outputs = new DeploymentOutputs(deployment);
         }
 
-        private sealed class DeploymentProperties : ILazyObject
-        {
-            private readonly DeploymentValue _Deployment;
-            private readonly JObject _Properties;
-            private readonly DeploymentOutputs _Outputs;
-
-            public DeploymentProperties(DeploymentValue deployment)
-            {
-                _Deployment = deployment;
-                _Deployment.Value.TryGetProperty(PROPERTY_PROPERTIES, out _Properties);
-                _Outputs = new DeploymentOutputs(deployment);
-            }
-
-            public bool TryProperty(string propertyName, out object value)
-            {
-                value = null;
-                if (StringComparer.OrdinalIgnoreCase.Equals(propertyName, PROPERTY_OUTPUTS))
-                    value = _Outputs;
-                else if (_Properties.TryGetProperty<JToken>(propertyName, out var token))
-                    value = token;
-
-                return value != null;
-            }
-        }
-
-        private sealed class DeploymentOutputs : JObject, ILazyObject
-        {
-            private readonly DeploymentValue _Deployment;
-
-            public DeploymentOutputs(DeploymentValue deployment)
-            {
-                _Deployment = deployment;
-            }
-
-            public bool TryProperty(string propertyName, out object value)
-            {
-                value = null;
-                if (!_Deployment._Outputs.TryGetValue(propertyName, out var lazy))
-                    return false;
-
-                value = lazy.GetValue();
-                return true;
-            }
-
-            public override JToken this[object key]
-            {
-                get
-                {
-                    return key is string propertyName && TryProperty(propertyName, out var o) ? JToken.FromObject(o) : default;
-                }
-                set
-                {
-
-                }
-            }
-        }
-
-        /// <inheritdoc/>
-        public JObject Value => _Value.Value;
-
-        /// <inheritdoc/>
-        public string Type => RESOURCETYPE_DEPLOYMENT;
-
-        /// <inheritdoc/>
-        public bool Existing => false;
-
-        /// <inheritdoc/>
-        public TemplateContext.CopyIndexState Copy { get; }
-
-        public ILazyObject Properties { get; }
-
-        public string Scope { get; }
-
-        public DeploymentScope DeploymentScope { get; }
-
-        public void AddOutput(string name, ILazyValue output)
-        {
-            _Outputs.Add(name, output);
-        }
-
-        internal bool TryOutput(string name, out object value)
+        public bool TryProperty(string propertyName, out object value)
         {
             value = null;
-            if (!_Outputs.TryGetValue(name, out var lazy))
+            if (StringComparer.OrdinalIgnoreCase.Equals(propertyName, PROPERTY_OUTPUTS))
+                value = _Outputs;
+            else if (_Properties.TryGetProperty<JToken>(propertyName, out var token))
+                value = token;
+
+            return value != null;
+        }
+    }
+
+    private sealed class DeploymentOutputs : JObject, ILazyObject
+    {
+        private readonly DeploymentValue _Deployment;
+
+        public DeploymentOutputs(DeploymentValue deployment)
+        {
+            _Deployment = deployment;
+        }
+
+        public bool TryProperty(string propertyName, out object value)
+        {
+            value = null;
+            if (!_Deployment._Outputs.TryGetValue(propertyName, out var lazy))
                 return false;
 
             value = lazy.GetValue();
             return true;
         }
 
-        public bool TryProperty(string propertyName, out object value)
+        public override JToken this[object key]
         {
-            value = null;
-            if (StringComparer.OrdinalIgnoreCase.Equals(propertyName, PROPERTY_PROPERTIES))
-                value = Properties;
-            else if (Value.TryGetProperty<JToken>(propertyName, out var token))
-                value = token;
+            get
+            {
+                return key is string propertyName && TryProperty(propertyName, out var o) ? JToken.FromObject(o) : default;
+            }
+            set
+            {
 
-            return value != null;
+            }
         }
+    }
+
+    /// <inheritdoc/>
+    public JObject Value => _Value.Value;
+
+    /// <inheritdoc/>
+    public string Type => RESOURCE_TYPE_DEPLOYMENT;
+
+    /// <inheritdoc/>
+    public bool Existing => false;
+
+    /// <inheritdoc/>
+    public TemplateContext.CopyIndexState Copy { get; }
+
+    public ILazyObject Properties { get; }
+
+    public string Scope { get; }
+
+    public DeploymentScope DeploymentScope { get; }
+
+    public void AddOutput(string name, ILazyValue output)
+    {
+        _Outputs.Add(name, output);
+    }
+
+    internal bool TryOutput(string name, out object value)
+    {
+        value = null;
+        if (!_Outputs.TryGetValue(name, out var lazy))
+            return false;
+
+        value = lazy.GetValue();
+        return true;
+    }
+
+    public bool TryProperty(string propertyName, out object value)
+    {
+        value = null;
+        if (StringComparer.OrdinalIgnoreCase.Equals(propertyName, PROPERTY_PROPERTIES))
+            value = Properties;
+        else if (Value.TryGetProperty<JToken>(propertyName, out var token))
+            value = token;
+
+        return value != null;
     }
 }
