+++ src/PSRule.Rules.Azure/Data/Template/FunctionErrorType.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Rules.Azure.Data.Template;
+
+/// <summary>
+/// The type of function error.
+/// </summary>
+public enum FunctionErrorType
+{
+    /// <summary>
+    /// An error cause by mismatching resource segments.
+    /// </summary>
+    MismatchingResourceSegments
+}
